feat: refactor photodiode handling and add task management for WebSocket and ESP-NOW communication

+++ include/photodiode.hpp
@@ -35,7 +35,7 @@ class Photodiode
     Photodiode();
     void begin();
     void update();
-    uint16_t convertToBits();
+    uint32_t convertToBits();
     float getDynamicThreshold();
     bool isBufferFull();
     bool isSampleBufferFull();
